backend/requirements.txt
Django==3.2.12
psycopg2-binary==2.9.3
gunicorn==20.1.0
djoser==2.0.5
djangorestframework==3.13.1 # https://github.com/sunscrapers/djoser/issues/541
markdown==3.3
django-filter==2.4.0
django_cors_headers==3.7.0
pytz==2021.3
sqlparse==0.4.2
pytest==3.6.0
pytest-django==3.2.1
django-debug-toolbar==1.9.1
Pillow==9.0.1
djangorestframework-simplejwt==5.0.0